+++ source/extensions/omni.isaac.wheeled_robots/omni/isaac/wheeled_robots/nodes/OgnDifferentialController.py
@@ -50,46 +50,70 @@ class OgnDifferentialController:
     @staticmethod
     def compute(db) -> bool:
         state = db.internal_state
-        if db.inputs.wheelRadius <= 0 or db.inputs.wheelDistance <= 0:
-            db.log_error("invalid wheel radius and distance")
+
+        try:
+            robot_param_bundle = db.inputs.robotParams
+            wheel_radius_input = (
+                -1
+            )  ## needed this in case incoming bundle doesn't have any attributes and miss the next for loop
+            wheel_distance_input = -1
+            for attr in robot_param_bundle.attributes:
+                if attr.name == "wheel_radius":
+                    wheel_radius_input = attr.value[0]
+                elif attr.name == "wheel_distance":
+                    wheel_distance_input = attr.value[0]
+
+            if wheel_radius_input <= 0 or wheel_distance_input <= 0:
+                db.log_warning("invalid wheel radius and distance")
+                return False
+
+            if (wheel_radius_input != state.wheel_radius) or (wheel_distance_input != state.wheel_distance):
+                state.wheel_radius = wheel_radius_input
+                state.wheel_distance = wheel_distance_input
+                state.initialized = False
+
+            vehicle_limit_bundle = db.inputs.vehicleLimits
+            for attr in vehicle_limit_bundle.attributes:
+                if attr.name == "max_linear_speed":
+                    max_linear_speed = attr.value[0]
+                    if max_linear_speed != state.max_linear_speed:
+                        state.max_linear_speed = max_linear_speed
+                        state.initialized = False
+                elif attr.name == "max_angular_speed":
+                    max_angular_speed = attr.value[0]
+                    if max_angular_speed != state.max_angular_speed:
+                        state.max_angular_speed = max_linear_speed
+                        state.initialized = False
+                elif attr.name == "max_wheel_speed":
+                    max_wheel_speed = attr.value[0]
+                    if max_wheel_speed != state.max_wheel_speed:
+                        state.max_wheel_speed = max_wheel_speed
+                        state.initialized = False
+
+            if not state.initialized:
+                state.initialize()
+
+            # send joint commands as a bundle out
+            command_bundle = db.outputs.jointCommands
+            command_bundle.clear()
+
+            joint_actions = state.forward(np.array([db.inputs.linearVelocity, db.inputs.angularVelocity]))
+            if joint_actions.joint_positions is not None:
+                position_attr = command_bundle.insert(
+                    (og.Type(og.BaseDataType.DOUBLE, array_depth=2), "joint_positions")
+                )
+                position_attr.value = joint_actions.joint_positions
+            if joint_actions.joint_velocities is not None:
+                velocity_attr = command_bundle.insert(
+                    (og.Type(og.BaseDataType.DOUBLE, array_depth=2), "joint_velocities")
+                )
+                velocity_attr.value = joint_actions.joint_velocities
+            if joint_actions.joint_efforts is not None:
+                effort_attr = command_bundle.insert((og.Type(og.BaseDataType.DOUBLE, array_depth=2), "joint_efforts"))
+                effort_attr.value = joint_actions.joint_efforts
+
+        except Exception as error:
+            db.log_warning(str(error))
             return False
-
-        if (db.inputs.wheelRadius != state.wheel_radius) or (db.inputs.wheelDistance != state.wheel_distance):
-            state.wheel_radius = db.inputs.wheelRadius
-            state.wheel_distance = db.inputs.wheelDistance
-            state.initialized = False
-
-        if db.inputs.maxLinearSpeed:
-            if db.inputs.max_linear_speed != state.maxLinearSpeed:
-                state.max_linear_speed = db.inputs.maxLinearSpeed
-                state.initialized = False
-
-        if db.inputs.maxAngularSpeed:
-            if db.inputs.max_angular_speed != state.maxAngularSpeed:
-                state.max_angular_speed = db.inputs.maxAngularSpeed
-                state.initialized = False
-
-        if db.inputs.maxWheelSpeed:
-            if db.inputs.max_wheel_speed != state.maxWheelSpeed:
-                state.max_wheel_speed = db.inputs.maxWheelSpeed
-                state.initialized = False
-
-        if not state.initialized:
-            state.initialize()
-
-        # send joint commands as a bundle out
-        command_bundle = db.outputs.jointCommands
-        command_bundle.clear()
-
-        joint_actions = state.forward(np.array([db.inputs.linearVelocity, db.inputs.angularVelocity]))
-        if joint_actions.joint_positions is not None:
-            position_attr = command_bundle.insert((og.Type(og.BaseDataType.DOUBLE, array_depth=2), "joint_positions"))
-            position_attr.value = joint_actions.joint_positions
-        if joint_actions.joint_velocities is not None:
-            velocity_attr = command_bundle.insert((og.Type(og.BaseDataType.DOUBLE, array_depth=2), "joint_velocities"))
-            velocity_attr.value = joint_actions.joint_velocities
-        if joint_actions.joint_efforts is not None:
-            effort_attr = command_bundle.insert((og.Type(og.BaseDataType.DOUBLE, array_depth=2), "joint_efforts"))
-            effort_attr.value = joint_actions.joint_efforts
 
         return True