+++ source/standalone_examples/api/omni.isaac.universal_robots/multiple_tasks.py
@@ -10,18 +10,17 @@ from omni.isaac.kit import SimulationApp
 
 simulation_app = SimulationApp({"headless": False})
 
-
+from omni.isaac.core.utils.nucleus import get_assets_root_path
 from omni.isaac.universal_robots.tasks import Stacking as UR10Stacking
 from omni.isaac.universal_robots.controllers import StackingController as UR10StackingController
 from omni.isaac.franka.tasks import Stacking as FrankaStacking
 from omni.isaac.dofbot.tasks import PickPlace
 from omni.isaac.franka.controllers import StackingController as FrankaStackingController
 from omni.isaac.core import World
-from omni.isaac.kaya import Kaya
-from omni.isaac.jetbot import Jetbot
+from omni.isaac.wheeled_robots.robots import WheeledRobot
 from omni.isaac.core.prims.xform_prim import XFormPrim
 from omni.isaac.wheeled_robots.controllers.holonomic_controller import HolonomicController
-from omni.isaac.jetbot.controllers import DifferentialController
+from omni.isaac.wheeled_robots.controllers.differential_controller import DifferentialController
 from omni.isaac.dofbot.controllers import PickPlaceController
 import numpy as np
 
@@ -35,8 +34,32 @@ tasks.append(UR10Stacking(name="task_1", offset=np.array([50, 50, 0])))
 my_world.add_task(tasks[-1])
 tasks.append(PickPlace(offset=np.array([0, -100, 0])))
 my_world.add_task(tasks[-1])
-my_kaya = my_world.scene.add(Kaya(prim_path="/World/Kaya", name="my_kaya", position=np.array([-100, 0, 0])))
-my_jetbot = my_world.scene.add(Jetbot(prim_path="/World/Jetbot", name="my_jetbot", position=np.array([-150, -150, 0])))
+assets_root_path = get_assets_root_path()
+if assets_root_path is None:
+    carb.log_error("Could not find Isaac Sim assets folder")
+kaya_asset_path = assets_root_path + "/Robots/Kaya/kaya.usd"
+my_kaya = my_world.scene.add(
+    WheeledRobot(
+        prim_path="/World/Kaya",
+        name="my_kaya",
+        wheel_dof_names=["axle_0_joint", "axle_1_joint", "axle_2_joint"],
+        create_robot=True,
+        usd_path=kaya_asset_path,
+        position=np.array([-100, 0, 0]),
+    )
+)
+jetbot_asset_path = assets_root_path + "/Robots/Jetbot/jetbot.usd"
+my_jetbot = my_world.scene.add(
+    WheeledRobot(
+        prim_path="/World/Jetbot",
+        name="my_jetbot",
+        wheel_dof_names=["left_wheel_joint", "right_wheel_joint"],
+        create_robot=True,
+        usd_path=jetbot_asset_path,
+        position=np.array([-150, -150, 0]),
+    )
+)
+
 my_world.reset()
 robots = []
 for i in range(num_of_tasks):
@@ -76,7 +99,7 @@ kaya_controller = HolonomicController(
     com_prim=XFormPrim("/World/kaya/base_link/control_offset"),
     angular_gain=1,
 )
-jetbot_controller = DifferentialController(name="simple_control")
+jetbot_controller = DifferentialController(name="simple_control", wheel_radius=3.0, wheel_base=11.25)
 pick_place_task_params = tasks[2].get_params()
 
 articulation_controllers = []