+++ source/extensions/omni.isaac.core_nodes/docs/CHANGELOG.md
@@ -2,12 +2,14 @@
 CHANGELOG
 **********
 
+
 [0.5.1] - 2022-05-03
 ========================
 
 Changed
 ---------
 - Output data types to vectord and quatd in Isaac Compute Odometry node
+- Articulation controller node takes bundles
 
 [0.5.0] - 2022-05-02
 ========================